2023-05/21 非同期通信

+++ app/controllers/user/favorites_controller.rb
@@ -1,15 +1,14 @@
 class User::FavoritesController < ApplicationController
   def create
-    illustration = Illustration.find(params[:illustration_id])
-    favorite = current_account.favorites.new(illustration_id: illustration.id)
-    favorite.save
-    redirect_to illustration_path(illustration)
+    @illustration = Illustration.find(params[:illustration_id])
+    favorite = current_account.favorites.find_or_create_by(illustration_id: @illustration.id)
+    render 'user/favorites/toggle'
   end
 
   def destroy
-    illustration = Illustration.find(params[:illustration_id])
-    favorite = current_account.favorites.find_by(illustration_id: illustration.id)
-    favorite.destroy
-    redirect_to illustration_path(illustration)
+    @illustration = Illustration.find(params[:illustration_id])
+    favorite = current_account.favorites.find_by(illustration_id: @illustration.id)
+    favorite.destroy if favorite
+    render 'user/favorites/toggle'
   end
 end
